+++ NEWS.md
@@ -14,6 +14,7 @@
 - Remove `f7ShowPreloader()` and `f7HidePreloader()` as long time deprecated.
 - `f7Menu()`, `f7MenuItem()`, `f7MenuDropdown()`, `f7MenuDropdownDivider()` and `updateF7MenuDropdown()` are totally removed from Framework7. We have no alternative to replace it.
 - `f7Icon()`: remove deprecated parameter `old`.
+- `f7SmartSelect()`: `maxlength` becomes `maxLength`. Typo from Framework7.
 
 ## Soft deprecation
 - `f7Accordion()`: